Method to identify cellular spatial neighbourhood from single cell spatial transcriptomics data.

hoodscanR implements a novel method to scan for cell neighbourhood from spatial transcriptomics data at single cell level, such as CosMx and MERFISH etc. hoodscanR takes the cellular position and cell type annotations as inputs, allowing cellular spatial neighbourhood analysis.

Details

Key neighborhood analysis functions include findNearCells, scanHoods, mergeByGroup, calcMetrics, clustByHood.

Key visualisation functions include plotTissue, plotHoodMat, plotColocal, plotProbDist.

See also

Author

Ning Liu liu.n@wehi.edu.au