Compute p-value for perplexity via permutation

Compute p-value for perplexity via permutation

Usage

perplexityPermute(spe, pm = NA, pm_cols = NA, n_perm = 1000)

Arguments

spe

A SpatialExperiment object.

pm

Optional. The probability matrix.

pm_cols

The colnames of probability matrix. This is requires for SpatialExperiment input. Assuming that the probability is stored in the colData.

n_perm

Integer number. The number of permutation. 1000 by default.

Value

A SpatialExperiment object. Calculated P-value and adjusted P-value are saved as columns in the colData of the SpatialExperiment object. P-value and adjusted P-value are calculated based on permutation test and Benjamini Hochberg correction.

Examples

data("spe_test")

spe <- readHoodData(spe, anno_col = "celltypes")

fnc <- findNearCells(spe, k = 100)

pm <- scanHoods(fnc$distance)
#> Tau is set to: 22747.4

pm2 <- mergeByGroup(pm, fnc$cells)

spe <- mergeHoodSpe(spe, pm2)

spe <- perplexityPermute(spe, pm_cols = colnames(pm2))
#> Generating permutation 1 of 1000Generating permutation 2 of 1000Generating permutation 3 of 1000Generating permutation 4 of 1000Generating permutation 5 of 1000Generating permutation 6 of 1000Generating permutation 7 of 1000Generating permutation 8 of 1000Generating permutation 9 of 1000Generating permutation 10 of 1000Generating permutation 11 of 1000Generating permutation 12 of 1000Generating permutation 13 of 1000Generating permutation 14 of 1000Generating permutation 15 of 1000Generating permutation 16 of 1000Generating permutation 17 of 1000Generating permutation 18 of 1000Generating permutation 19 of 1000Generating permutation 20 of 1000Generating permutation 21 of 1000Generating permutation 22 of 1000Generating permutation 23 of 1000Generating permutation 24 of 1000Generating permutation 25 of 1000Generating permutation 26 of 1000Generating permutation 27 of 1000Generating permutation 28 of 1000Generating permutation 29 of 1000Generating permutation 30 of 1000Generating permutation 31 of 1000Generating permutation 32 of 1000Generating permutation 33 of 1000Generating permutation 34 of 1000Generating permutation 35 of 1000Generating permutation 36 of 1000Generating permutation 37 of 1000Generating permutation 38 of 1000Generating permutation 39 of 1000Generating permutation 40 of 1000Generating permutation 41 of 1000Generating permutation 42 of 1000Generating permutation 43 of 1000Generating permutation 44 of 1000Generating permutation 45 of 1000Generating permutation 46 of 1000Generating permutation 47 of 1000Generating permutation 48 of 1000Generating permutation 49 of 1000Generating permutation 50 of 1000Generating permutation 51 of 1000Generating permutation 52 of 1000Generating permutation 53 of 1000Generating permutation 54 of 1000Generating permutation 55 of 1000Generating permutation 56 of 1000Generating permutation 57 of 1000Generating permutation 58 of 1000Generating permutation 59 of 1000Generating permutation 60 of 1000Generating permutation 61 of 1000Generating permutation 62 of 1000Generating permutation 63 of 1000Generating permutation 64 of 1000Generating permutation 65 of 1000Generating permutation 66 of 1000Generating permutation 67 of 1000Generating permutation 68 of 1000Generating permutation 69 of 1000Generating permutation 70 of 1000Generating permutation 71 of 1000Generating permutation 72 of 1000Generating permutation 73 of 1000Generating permutation 74 of 1000Generating permutation 75 of 1000Generating permutation 76 of 1000Generating permutation 77 of 1000Generating permutation 78 of 1000Generating permutation 79 of 1000Generating permutation 80 of 1000Generating permutation 81 of 1000Generating permutation 82 of 1000Generating permutation 83 of 1000Generating permutation 84 of 1000Generating permutation 85 of 1000Generating permutation 86 of 1000Generating permutation 87 of 1000Generating permutation 88 of 1000Generating permutation 89 of 1000Generating permutation 90 of 1000Generating permutation 91 of 1000Generating permutation 92 of 1000Generating permutation 93 of 1000Generating permutation 94 of 1000Generating permutation 95 of 1000Generating permutation 96 of 1000Generating permutation 97 of 1000Generating permutation 98 of 1000Generating permutation 99 of 1000Generating permutation 100 of 1000Generating permutation 101 of 1000Generating permutation 102 of 1000Generating permutation 103 of 1000Generating permutation 104 of 1000Generating permutation 105 of 1000Generating permutation 106 of 1000Generating permutation 107 of 1000Generating permutation 108 of 1000Generating permutation 109 of 1000Generating permutation 110 of 1000Generating permutation 111 of 1000Generating permutation 112 of 1000Generating permutation 113 of 1000Generating permutation 114 of 1000Generating permutation 115 of 1000Generating permutation 116 of 1000Generating permutation 117 of 1000Generating permutation 118 of 1000Generating permutation 119 of 1000Generating permutation 120 of 1000Generating permutation 121 of 1000Generating permutation 122 of 1000Generating permutation 123 of 1000Generating permutation 124 of 1000Generating permutation 125 of 1000Generating permutation 126 of 1000Generating permutation 127 of 1000Generating permutation 128 of 1000Generating permutation 129 of 1000Generating permutation 130 of 1000Generating permutation 131 of 1000Generating permutation 132 of 1000Generating permutation 133 of 1000Generating permutation 134 of 1000Generating permutation 135 of 1000Generating permutation 136 of 1000Generating permutation 137 of 1000Generating permutation 138 of 1000Generating permutation 139 of 1000Generating permutation 140 of 1000Generating permutation 141 of 1000Generating permutation 142 of 1000Generating permutation 143 of 1000Generating permutation 144 of 1000Generating permutation 145 of 1000Generating permutation 146 of 1000Generating permutation 147 of 1000Generating permutation 148 of 1000Generating permutation 149 of 1000Generating permutation 150 of 1000Generating permutation 151 of 1000Generating permutation 152 of 1000Generating permutation 153 of 1000Generating permutation 154 of 1000Generating permutation 155 of 1000Generating permutation 156 of 1000Generating permutation 157 of 1000Generating permutation 158 of 1000Generating permutation 159 of 1000Generating permutation 160 of 1000Generating permutation 161 of 1000Generating permutation 162 of 1000Generating permutation 163 of 1000Generating permutation 164 of 1000Generating permutation 165 of 1000Generating permutation 166 of 1000Generating permutation 167 of 1000Generating permutation 168 of 1000Generating permutation 169 of 1000Generating permutation 170 of 1000Generating permutation 171 of 1000Generating permutation 172 of 1000Generating permutation 173 of 1000Generating permutation 174 of 1000Generating permutation 175 of 1000Generating permutation 176 of 1000Generating permutation 177 of 1000Generating permutation 178 of 1000Generating permutation 179 of 1000Generating permutation 180 of 1000Generating permutation 181 of 1000Generating permutation 182 of 1000Generating permutation 183 of 1000Generating permutation 184 of 1000Generating permutation 185 of 1000Generating permutation 186 of 1000Generating permutation 187 of 1000Generating permutation 188 of 1000Generating permutation 189 of 1000Generating permutation 190 of 1000Generating permutation 191 of 1000Generating permutation 192 of 1000Generating permutation 193 of 1000Generating permutation 194 of 1000Generating permutation 195 of 1000Generating permutation 196 of 1000Generating permutation 197 of 1000Generating permutation 198 of 1000Generating permutation 199 of 1000Generating permutation 200 of 1000Generating permutation 201 of 1000Generating permutation 202 of 1000Generating permutation 203 of 1000Generating permutation 204 of 1000Generating permutation 205 of 1000Generating permutation 206 of 1000Generating permutation 207 of 1000Generating permutation 208 of 1000Generating permutation 209 of 1000Generating permutation 210 of 1000Generating permutation 211 of 1000Generating permutation 212 of 1000Generating permutation 213 of 1000Generating permutation 214 of 1000Generating permutation 215 of 1000Generating permutation 216 of 1000Generating permutation 217 of 1000Generating permutation 218 of 1000Generating permutation 219 of 1000Generating permutation 220 of 1000Generating permutation 221 of 1000Generating permutation 222 of 1000Generating permutation 223 of 1000Generating permutation 224 of 1000Generating permutation 225 of 1000Generating permutation 226 of 1000Generating permutation 227 of 1000Generating permutation 228 of 1000Generating permutation 229 of 1000Generating permutation 230 of 1000Generating permutation 231 of 1000Generating permutation 232 of 1000Generating permutation 233 of 1000Generating permutation 234 of 1000Generating permutation 235 of 1000Generating permutation 236 of 1000Generating permutation 237 of 1000Generating permutation 238 of 1000Generating permutation 239 of 1000Generating permutation 240 of 1000Generating permutation 241 of 1000Generating permutation 242 of 1000Generating permutation 243 of 1000Generating permutation 244 of 1000Generating permutation 245 of 1000Generating permutation 246 of 1000Generating permutation 247 of 1000Generating permutation 248 of 1000Generating permutation 249 of 1000Generating permutation 250 of 1000Generating permutation 251 of 1000Generating permutation 252 of 1000Generating permutation 253 of 1000Generating permutation 254 of 1000Generating permutation 255 of 1000Generating permutation 256 of 1000Generating permutation 257 of 1000Generating permutation 258 of 1000Generating permutation 259 of 1000Generating permutation 260 of 1000Generating permutation 261 of 1000Generating permutation 262 of 1000Generating permutation 263 of 1000Generating permutation 264 of 1000Generating permutation 265 of 1000Generating permutation 266 of 1000Generating permutation 267 of 1000Generating permutation 268 of 1000Generating permutation 269 of 1000Generating permutation 270 of 1000Generating permutation 271 of 1000Generating permutation 272 of 1000Generating permutation 273 of 1000Generating permutation 274 of 1000Generating permutation 275 of 1000Generating permutation 276 of 1000Generating permutation 277 of 1000Generating permutation 278 of 1000Generating permutation 279 of 1000Generating permutation 280 of 1000Generating permutation 281 of 1000Generating permutation 282 of 1000Generating permutation 283 of 1000Generating permutation 284 of 1000Generating permutation 285 of 1000Generating permutation 286 of 1000Generating permutation 287 of 1000Generating permutation 288 of 1000Generating permutation 289 of 1000Generating permutation 290 of 1000Generating permutation 291 of 1000Generating permutation 292 of 1000Generating permutation 293 of 1000Generating permutation 294 of 1000Generating permutation 295 of 1000Generating permutation 296 of 1000Generating permutation 297 of 1000Generating permutation 298 of 1000Generating permutation 299 of 1000Generating permutation 300 of 1000Generating permutation 301 of 1000Generating permutation 302 of 1000Generating permutation 303 of 1000Generating permutation 304 of 1000Generating permutation 305 of 1000Generating permutation 306 of 1000Generating permutation 307 of 1000Generating permutation 308 of 1000Generating permutation 309 of 1000Generating permutation 310 of 1000Generating permutation 311 of 1000Generating permutation 312 of 1000Generating permutation 313 of 1000Generating permutation 314 of 1000Generating permutation 315 of 1000Generating permutation 316 of 1000Generating permutation 317 of 1000Generating permutation 318 of 1000Generating permutation 319 of 1000Generating permutation 320 of 1000Generating permutation 321 of 1000Generating permutation 322 of 1000Generating permutation 323 of 1000Generating permutation 324 of 1000Generating permutation 325 of 1000Generating permutation 326 of 1000Generating permutation 327 of 1000Generating permutation 328 of 1000Generating permutation 329 of 1000Generating permutation 330 of 1000Generating permutation 331 of 1000Generating permutation 332 of 1000Generating permutation 333 of 1000Generating permutation 334 of 1000Generating permutation 335 of 1000Generating permutation 336 of 1000Generating permutation 337 of 1000Generating permutation 338 of 1000Generating permutation 339 of 1000Generating permutation 340 of 1000Generating permutation 341 of 1000Generating permutation 342 of 1000Generating permutation 343 of 1000Generating permutation 344 of 1000Generating permutation 345 of 1000Generating permutation 346 of 1000Generating permutation 347 of 1000Generating permutation 348 of 1000Generating permutation 349 of 1000Generating permutation 350 of 1000Generating permutation 351 of 1000Generating permutation 352 of 1000Generating permutation 353 of 1000Generating permutation 354 of 1000Generating permutation 355 of 1000Generating permutation 356 of 1000Generating permutation 357 of 1000Generating permutation 358 of 1000Generating permutation 359 of 1000Generating permutation 360 of 1000Generating permutation 361 of 1000Generating permutation 362 of 1000Generating permutation 363 of 1000Generating permutation 364 of 1000Generating permutation 365 of 1000Generating permutation 366 of 1000Generating permutation 367 of 1000Generating permutation 368 of 1000Generating permutation 369 of 1000Generating permutation 370 of 1000Generating permutation 371 of 1000Generating permutation 372 of 1000Generating permutation 373 of 1000Generating permutation 374 of 1000Generating permutation 375 of 1000Generating permutation 376 of 1000Generating permutation 377 of 1000Generating permutation 378 of 1000Generating permutation 379 of 1000Generating permutation 380 of 1000Generating permutation 381 of 1000Generating permutation 382 of 1000Generating permutation 383 of 1000Generating permutation 384 of 1000Generating permutation 385 of 1000Generating permutation 386 of 1000Generating permutation 387 of 1000Generating permutation 388 of 1000Generating permutation 389 of 1000Generating permutation 390 of 1000Generating permutation 391 of 1000Generating permutation 392 of 1000Generating permutation 393 of 1000Generating permutation 394 of 1000Generating permutation 395 of 1000Generating permutation 396 of 1000Generating permutation 397 of 1000Generating permutation 398 of 1000Generating permutation 399 of 1000Generating permutation 400 of 1000Generating permutation 401 of 1000Generating permutation 402 of 1000Generating permutation 403 of 1000Generating permutation 404 of 1000Generating permutation 405 of 1000Generating permutation 406 of 1000Generating permutation 407 of 1000Generating permutation 408 of 1000Generating permutation 409 of 1000Generating permutation 410 of 1000Generating permutation 411 of 1000Generating permutation 412 of 1000Generating permutation 413 of 1000Generating permutation 414 of 1000Generating permutation 415 of 1000Generating permutation 416 of 1000Generating permutation 417 of 1000Generating permutation 418 of 1000Generating permutation 419 of 1000Generating permutation 420 of 1000Generating permutation 421 of 1000Generating permutation 422 of 1000Generating permutation 423 of 1000Generating permutation 424 of 1000Generating permutation 425 of 1000Generating permutation 426 of 1000Generating permutation 427 of 1000Generating permutation 428 of 1000Generating permutation 429 of 1000Generating permutation 430 of 1000Generating permutation 431 of 1000Generating permutation 432 of 1000Generating permutation 433 of 1000Generating permutation 434 of 1000Generating permutation 435 of 1000Generating permutation 436 of 1000Generating permutation 437 of 1000Generating permutation 438 of 1000Generating permutation 439 of 1000Generating permutation 440 of 1000Generating permutation 441 of 1000Generating permutation 442 of 1000Generating permutation 443 of 1000Generating permutation 444 of 1000Generating permutation 445 of 1000Generating permutation 446 of 1000Generating permutation 447 of 1000Generating permutation 448 of 1000Generating permutation 449 of 1000Generating permutation 450 of 1000Generating permutation 451 of 1000Generating permutation 452 of 1000Generating permutation 453 of 1000Generating permutation 454 of 1000Generating permutation 455 of 1000Generating permutation 456 of 1000Generating permutation 457 of 1000Generating permutation 458 of 1000Generating permutation 459 of 1000Generating permutation 460 of 1000Generating permutation 461 of 1000Generating permutation 462 of 1000Generating permutation 463 of 1000Generating permutation 464 of 1000Generating permutation 465 of 1000Generating permutation 466 of 1000Generating permutation 467 of 1000Generating permutation 468 of 1000Generating permutation 469 of 1000Generating permutation 470 of 1000Generating permutation 471 of 1000Generating permutation 472 of 1000Generating permutation 473 of 1000Generating permutation 474 of 1000Generating permutation 475 of 1000Generating permutation 476 of 1000Generating permutation 477 of 1000Generating permutation 478 of 1000Generating permutation 479 of 1000Generating permutation 480 of 1000Generating permutation 481 of 1000Generating permutation 482 of 1000Generating permutation 483 of 1000Generating permutation 484 of 1000Generating permutation 485 of 1000Generating permutation 486 of 1000Generating permutation 487 of 1000Generating permutation 488 of 1000Generating permutation 489 of 1000Generating permutation 490 of 1000Generating permutation 491 of 1000Generating permutation 492 of 1000Generating permutation 493 of 1000Generating permutation 494 of 1000Generating permutation 495 of 1000Generating permutation 496 of 1000Generating permutation 497 of 1000Generating permutation 498 of 1000Generating permutation 499 of 1000Generating permutation 500 of 1000Generating permutation 501 of 1000Generating permutation 502 of 1000Generating permutation 503 of 1000Generating permutation 504 of 1000Generating permutation 505 of 1000Generating permutation 506 of 1000Generating permutation 507 of 1000Generating permutation 508 of 1000Generating permutation 509 of 1000Generating permutation 510 of 1000Generating permutation 511 of 1000Generating permutation 512 of 1000Generating permutation 513 of 1000Generating permutation 514 of 1000Generating permutation 515 of 1000Generating permutation 516 of 1000Generating permutation 517 of 1000Generating permutation 518 of 1000Generating permutation 519 of 1000Generating permutation 520 of 1000Generating permutation 521 of 1000Generating permutation 522 of 1000Generating permutation 523 of 1000Generating permutation 524 of 1000Generating permutation 525 of 1000Generating permutation 526 of 1000Generating permutation 527 of 1000Generating permutation 528 of 1000Generating permutation 529 of 1000Generating permutation 530 of 1000Generating permutation 531 of 1000Generating permutation 532 of 1000Generating permutation 533 of 1000Generating permutation 534 of 1000Generating permutation 535 of 1000Generating permutation 536 of 1000Generating permutation 537 of 1000Generating permutation 538 of 1000Generating permutation 539 of 1000Generating permutation 540 of 1000Generating permutation 541 of 1000Generating permutation 542 of 1000Generating permutation 543 of 1000Generating permutation 544 of 1000Generating permutation 545 of 1000Generating permutation 546 of 1000Generating permutation 547 of 1000Generating permutation 548 of 1000Generating permutation 549 of 1000Generating permutation 550 of 1000Generating permutation 551 of 1000Generating permutation 552 of 1000Generating permutation 553 of 1000Generating permutation 554 of 1000Generating permutation 555 of 1000Generating permutation 556 of 1000Generating permutation 557 of 1000Generating permutation 558 of 1000Generating permutation 559 of 1000Generating permutation 560 of 1000Generating permutation 561 of 1000Generating permutation 562 of 1000Generating permutation 563 of 1000Generating permutation 564 of 1000Generating permutation 565 of 1000Generating permutation 566 of 1000Generating permutation 567 of 1000Generating permutation 568 of 1000Generating permutation 569 of 1000Generating permutation 570 of 1000Generating permutation 571 of 1000Generating permutation 572 of 1000Generating permutation 573 of 1000Generating permutation 574 of 1000Generating permutation 575 of 1000Generating permutation 576 of 1000Generating permutation 577 of 1000Generating permutation 578 of 1000Generating permutation 579 of 1000Generating permutation 580 of 1000Generating permutation 581 of 1000Generating permutation 582 of 1000Generating permutation 583 of 1000Generating permutation 584 of 1000Generating permutation 585 of 1000Generating permutation 586 of 1000Generating permutation 587 of 1000Generating permutation 588 of 1000Generating permutation 589 of 1000Generating permutation 590 of 1000Generating permutation 591 of 1000Generating permutation 592 of 1000Generating permutation 593 of 1000Generating permutation 594 of 1000Generating permutation 595 of 1000Generating permutation 596 of 1000Generating permutation 597 of 1000Generating permutation 598 of 1000Generating permutation 599 of 1000Generating permutation 600 of 1000Generating permutation 601 of 1000Generating permutation 602 of 1000Generating permutation 603 of 1000Generating permutation 604 of 1000Generating permutation 605 of 1000Generating permutation 606 of 1000Generating permutation 607 of 1000Generating permutation 608 of 1000Generating permutation 609 of 1000Generating permutation 610 of 1000Generating permutation 611 of 1000Generating permutation 612 of 1000Generating permutation 613 of 1000Generating permutation 614 of 1000Generating permutation 615 of 1000Generating permutation 616 of 1000Generating permutation 617 of 1000Generating permutation 618 of 1000Generating permutation 619 of 1000Generating permutation 620 of 1000Generating permutation 621 of 1000Generating permutation 622 of 1000Generating permutation 623 of 1000Generating permutation 624 of 1000Generating permutation 625 of 1000Generating permutation 626 of 1000Generating permutation 627 of 1000Generating permutation 628 of 1000Generating permutation 629 of 1000Generating permutation 630 of 1000Generating permutation 631 of 1000Generating permutation 632 of 1000Generating permutation 633 of 1000Generating permutation 634 of 1000Generating permutation 635 of 1000Generating permutation 636 of 1000Generating permutation 637 of 1000Generating permutation 638 of 1000Generating permutation 639 of 1000Generating permutation 640 of 1000Generating permutation 641 of 1000Generating permutation 642 of 1000Generating permutation 643 of 1000Generating permutation 644 of 1000Generating permutation 645 of 1000Generating permutation 646 of 1000Generating permutation 647 of 1000Generating permutation 648 of 1000Generating permutation 649 of 1000Generating permutation 650 of 1000Generating permutation 651 of 1000Generating permutation 652 of 1000Generating permutation 653 of 1000Generating permutation 654 of 1000Generating permutation 655 of 1000Generating permutation 656 of 1000Generating permutation 657 of 1000Generating permutation 658 of 1000Generating permutation 659 of 1000Generating permutation 660 of 1000Generating permutation 661 of 1000Generating permutation 662 of 1000Generating permutation 663 of 1000Generating permutation 664 of 1000Generating permutation 665 of 1000Generating permutation 666 of 1000Generating permutation 667 of 1000Generating permutation 668 of 1000Generating permutation 669 of 1000Generating permutation 670 of 1000Generating permutation 671 of 1000Generating permutation 672 of 1000Generating permutation 673 of 1000Generating permutation 674 of 1000Generating permutation 675 of 1000Generating permutation 676 of 1000Generating permutation 677 of 1000Generating permutation 678 of 1000Generating permutation 679 of 1000Generating permutation 680 of 1000Generating permutation 681 of 1000Generating permutation 682 of 1000Generating permutation 683 of 1000Generating permutation 684 of 1000Generating permutation 685 of 1000Generating permutation 686 of 1000Generating permutation 687 of 1000Generating permutation 688 of 1000Generating permutation 689 of 1000Generating permutation 690 of 1000Generating permutation 691 of 1000Generating permutation 692 of 1000Generating permutation 693 of 1000Generating permutation 694 of 1000Generating permutation 695 of 1000Generating permutation 696 of 1000Generating permutation 697 of 1000Generating permutation 698 of 1000Generating permutation 699 of 1000Generating permutation 700 of 1000Generating permutation 701 of 1000Generating permutation 702 of 1000Generating permutation 703 of 1000Generating permutation 704 of 1000Generating permutation 705 of 1000Generating permutation 706 of 1000Generating permutation 707 of 1000Generating permutation 708 of 1000Generating permutation 709 of 1000Generating permutation 710 of 1000Generating permutation 711 of 1000Generating permutation 712 of 1000Generating permutation 713 of 1000Generating permutation 714 of 1000Generating permutation 715 of 1000Generating permutation 716 of 1000Generating permutation 717 of 1000Generating permutation 718 of 1000Generating permutation 719 of 1000Generating permutation 720 of 1000Generating permutation 721 of 1000Generating permutation 722 of 1000Generating permutation 723 of 1000Generating permutation 724 of 1000Generating permutation 725 of 1000Generating permutation 726 of 1000Generating permutation 727 of 1000Generating permutation 728 of 1000Generating permutation 729 of 1000Generating permutation 730 of 1000Generating permutation 731 of 1000Generating permutation 732 of 1000Generating permutation 733 of 1000Generating permutation 734 of 1000Generating permutation 735 of 1000Generating permutation 736 of 1000Generating permutation 737 of 1000Generating permutation 738 of 1000Generating permutation 739 of 1000Generating permutation 740 of 1000Generating permutation 741 of 1000Generating permutation 742 of 1000Generating permutation 743 of 1000Generating permutation 744 of 1000Generating permutation 745 of 1000Generating permutation 746 of 1000Generating permutation 747 of 1000Generating permutation 748 of 1000Generating permutation 749 of 1000Generating permutation 750 of 1000Generating permutation 751 of 1000Generating permutation 752 of 1000Generating permutation 753 of 1000Generating permutation 754 of 1000Generating permutation 755 of 1000Generating permutation 756 of 1000Generating permutation 757 of 1000Generating permutation 758 of 1000Generating permutation 759 of 1000Generating permutation 760 of 1000Generating permutation 761 of 1000Generating permutation 762 of 1000Generating permutation 763 of 1000Generating permutation 764 of 1000Generating permutation 765 of 1000Generating permutation 766 of 1000Generating permutation 767 of 1000Generating permutation 768 of 1000Generating permutation 769 of 1000Generating permutation 770 of 1000Generating permutation 771 of 1000Generating permutation 772 of 1000Generating permutation 773 of 1000Generating permutation 774 of 1000Generating permutation 775 of 1000Generating permutation 776 of 1000Generating permutation 777 of 1000Generating permutation 778 of 1000Generating permutation 779 of 1000Generating permutation 780 of 1000Generating permutation 781 of 1000Generating permutation 782 of 1000Generating permutation 783 of 1000Generating permutation 784 of 1000Generating permutation 785 of 1000Generating permutation 786 of 1000Generating permutation 787 of 1000Generating permutation 788 of 1000Generating permutation 789 of 1000Generating permutation 790 of 1000Generating permutation 791 of 1000Generating permutation 792 of 1000Generating permutation 793 of 1000Generating permutation 794 of 1000Generating permutation 795 of 1000Generating permutation 796 of 1000Generating permutation 797 of 1000Generating permutation 798 of 1000Generating permutation 799 of 1000Generating permutation 800 of 1000Generating permutation 801 of 1000Generating permutation 802 of 1000Generating permutation 803 of 1000Generating permutation 804 of 1000Generating permutation 805 of 1000Generating permutation 806 of 1000Generating permutation 807 of 1000Generating permutation 808 of 1000Generating permutation 809 of 1000Generating permutation 810 of 1000Generating permutation 811 of 1000Generating permutation 812 of 1000Generating permutation 813 of 1000Generating permutation 814 of 1000Generating permutation 815 of 1000Generating permutation 816 of 1000Generating permutation 817 of 1000Generating permutation 818 of 1000Generating permutation 819 of 1000Generating permutation 820 of 1000Generating permutation 821 of 1000Generating permutation 822 of 1000Generating permutation 823 of 1000Generating permutation 824 of 1000Generating permutation 825 of 1000Generating permutation 826 of 1000Generating permutation 827 of 1000Generating permutation 828 of 1000Generating permutation 829 of 1000Generating permutation 830 of 1000Generating permutation 831 of 1000Generating permutation 832 of 1000Generating permutation 833 of 1000Generating permutation 834 of 1000Generating permutation 835 of 1000Generating permutation 836 of 1000Generating permutation 837 of 1000Generating permutation 838 of 1000Generating permutation 839 of 1000Generating permutation 840 of 1000Generating permutation 841 of 1000Generating permutation 842 of 1000Generating permutation 843 of 1000Generating permutation 844 of 1000Generating permutation 845 of 1000Generating permutation 846 of 1000Generating permutation 847 of 1000Generating permutation 848 of 1000Generating permutation 849 of 1000Generating permutation 850 of 1000Generating permutation 851 of 1000Generating permutation 852 of 1000Generating permutation 853 of 1000Generating permutation 854 of 1000Generating permutation 855 of 1000Generating permutation 856 of 1000Generating permutation 857 of 1000Generating permutation 858 of 1000Generating permutation 859 of 1000Generating permutation 860 of 1000Generating permutation 861 of 1000Generating permutation 862 of 1000Generating permutation 863 of 1000Generating permutation 864 of 1000Generating permutation 865 of 1000Generating permutation 866 of 1000Generating permutation 867 of 1000Generating permutation 868 of 1000Generating permutation 869 of 1000Generating permutation 870 of 1000Generating permutation 871 of 1000Generating permutation 872 of 1000Generating permutation 873 of 1000Generating permutation 874 of 1000Generating permutation 875 of 1000Generating permutation 876 of 1000Generating permutation 877 of 1000Generating permutation 878 of 1000Generating permutation 879 of 1000Generating permutation 880 of 1000Generating permutation 881 of 1000Generating permutation 882 of 1000Generating permutation 883 of 1000Generating permutation 884 of 1000Generating permutation 885 of 1000Generating permutation 886 of 1000Generating permutation 887 of 1000Generating permutation 888 of 1000Generating permutation 889 of 1000Generating permutation 890 of 1000Generating permutation 891 of 1000Generating permutation 892 of 1000Generating permutation 893 of 1000Generating permutation 894 of 1000Generating permutation 895 of 1000Generating permutation 896 of 1000Generating permutation 897 of 1000Generating permutation 898 of 1000Generating permutation 899 of 1000Generating permutation 900 of 1000Generating permutation 901 of 1000Generating permutation 902 of 1000Generating permutation 903 of 1000Generating permutation 904 of 1000Generating permutation 905 of 1000Generating permutation 906 of 1000Generating permutation 907 of 1000Generating permutation 908 of 1000Generating permutation 909 of 1000Generating permutation 910 of 1000Generating permutation 911 of 1000Generating permutation 912 of 1000Generating permutation 913 of 1000Generating permutation 914 of 1000Generating permutation 915 of 1000Generating permutation 916 of 1000Generating permutation 917 of 1000Generating permutation 918 of 1000Generating permutation 919 of 1000Generating permutation 920 of 1000Generating permutation 921 of 1000Generating permutation 922 of 1000Generating permutation 923 of 1000Generating permutation 924 of 1000Generating permutation 925 of 1000Generating permutation 926 of 1000Generating permutation 927 of 1000Generating permutation 928 of 1000Generating permutation 929 of 1000Generating permutation 930 of 1000Generating permutation 931 of 1000Generating permutation 932 of 1000Generating permutation 933 of 1000Generating permutation 934 of 1000Generating permutation 935 of 1000Generating permutation 936 of 1000Generating permutation 937 of 1000Generating permutation 938 of 1000Generating permutation 939 of 1000Generating permutation 940 of 1000Generating permutation 941 of 1000Generating permutation 942 of 1000Generating permutation 943 of 1000Generating permutation 944 of 1000Generating permutation 945 of 1000Generating permutation 946 of 1000Generating permutation 947 of 1000Generating permutation 948 of 1000Generating permutation 949 of 1000Generating permutation 950 of 1000Generating permutation 951 of 1000Generating permutation 952 of 1000Generating permutation 953 of 1000Generating permutation 954 of 1000Generating permutation 955 of 1000Generating permutation 956 of 1000Generating permutation 957 of 1000Generating permutation 958 of 1000Generating permutation 959 of 1000Generating permutation 960 of 1000Generating permutation 961 of 1000Generating permutation 962 of 1000Generating permutation 963 of 1000Generating permutation 964 of 1000Generating permutation 965 of 1000Generating permutation 966 of 1000Generating permutation 967 of 1000Generating permutation 968 of 1000Generating permutation 969 of 1000Generating permutation 970 of 1000Generating permutation 971 of 1000Generating permutation 972 of 1000Generating permutation 973 of 1000Generating permutation 974 of 1000Generating permutation 975 of 1000Generating permutation 976 of 1000Generating permutation 977 of 1000Generating permutation 978 of 1000Generating permutation 979 of 1000Generating permutation 980 of 1000Generating permutation 981 of 1000Generating permutation 982 of 1000Generating permutation 983 of 1000Generating permutation 984 of 1000Generating permutation 985 of 1000Generating permutation 986 of 1000Generating permutation 987 of 1000Generating permutation 988 of 1000Generating permutation 989 of 1000Generating permutation 990 of 1000Generating permutation 991 of 1000Generating permutation 992 of 1000Generating permutation 993 of 1000Generating permutation 994 of 1000Generating permutation 995 of 1000Generating permutation 996 of 1000Generating permutation 997 of 1000Generating permutation 998 of 1000Generating permutation 999 of 1000Generating permutation 1000 of 1000